Fix background rectangle coordinates
marginX()/marginY() return unsigned values, so negating them
caused unsigned wraparound and passed huge coordinates to
cairo_rectangle().

Cast them to signed/double before negation.

+++ src/virtualkeyboard.cpp
@@ -233,8 +233,11 @@ void VirtualKeyboard::paint(cairo_t *cr, unsigned int offsetX, unsigned int offs
 void VirtualKeyboard::paintBackground(cairo_t *cr) {
     auto [keyboardWidth, keyboardHeight] = size();
     cairo_set_source_rgb(cr, 0.95, 0.95, 0.95);
-    cairo_rectangle(cr, -marginX(), -marginY(),
-        keyboardWidth  + 2 * marginX(), keyboardHeight + 2 * marginY());
+    cairo_rectangle(cr,
+                    -static_cast<double>(marginX()),
+                    -static_cast<double>(marginY()),
+                    keyboardWidth  + 2 * marginX(),
+                    keyboardHeight + 2 * marginY());
     cairo_fill(cr);
 }
